Trait cleopatra_cairo::vm::runners::builtin_runner::BuiltinRunner

pub trait BuiltinRunner {
    fn initialize_segments(
        &mut self,
        segments: &mut MemorySegmentManager,
        memory: &mut Memory
    );
    fn initial_stack(&self) -> Result<Vec<MaybeRelocatable>, RunnerError>;
    fn base(&self) -> Option<Relocatable>;
    fn add_validation_rule(&self, memory: &mut Memory);
    fn deduce_memory_cell(
        &mut self,
        address: &MaybeRelocatable,
        memory: &Memory
    ) -> Result<Option<MaybeRelocatable>, RunnerError>;
}

Required Methods

fn initialize_segments(
&mut self,
segments: &mut MemorySegmentManager,
memory: &mut Memory
)

Creates the necessary segments for the builtin in the MemorySegmentManager and stores the first address on the builtin’s base

fn initial_stack(&self) -> Result<Vec<MaybeRelocatable>, RunnerError>

fn base(&self) -> Option<Relocatable>

Returns the builtin’s base

fn add_validation_rule(&self, memory: &mut Memory)

fn deduce_memory_cell(
&mut self,
address: &MaybeRelocatable,
memory: &Memory
) -> Result<Option<MaybeRelocatable>, RunnerError>